Further tidying-up of codebase.
Unfortunately, the way the code is used has had to change a little.

Some code in the "plotygons" module was boilerplate, or reproduced
functionality which was already present. That's been streamlined.

I've added a matplotlib-based demo. It's not the prettiest possible,
but it's a good test.

I've stripped out all the matplotlib references from tiling2.py: it is
possible to want to use a Tiling2 object without wanting to use
matplotlib, and so it's inappropriate to have to import matplotlib
(which is slow) just to get at tiling2. The downside is that this
means that Tiling2 objects can no longer have matplotlib methods.